Created a script consolidating all input images and generating an image label index
